cloudFunctions/gcs-file-fanout/requirements.txt
functions-framework
google-cloud-pubsub
google-cloud-tasks